+++ packages/expo-widgets/ios/WidgetsModule.swift
@@ -86,8 +86,7 @@ public final class WidgetsModule: Module {
         LiveActivityFactory(name: name, layout: layout)
       }
 
-      Function("start") { (liveActivity: LiveActivityFactory, props: String?, url: URL?, staleDateMs: Double?) in
-        let staleDate = staleDateMs.map { Date(timeIntervalSince1970: $0 / 1000.0) }
+      Function("start") { (liveActivity: LiveActivityFactory, props: String?, url: URL?, staleDate: Date?) in
         return try liveActivity.start(props: props, url: url, staleDate: staleDate)
       }
 
@@ -97,8 +96,7 @@ public final class WidgetsModule: Module {
     }
 
     Class("LiveActivity", LiveActivity.self) {
-      AsyncFunction("update") { (instance: LiveActivity, props: String?, staleDateMs: Double?) in
-        let staleDate = staleDateMs.map { Date(timeIntervalSince1970: $0 / 1000.0) }
+      AsyncFunction("update") { (instance: LiveActivity, props: String?, staleDate: Date?) in
         try await instance.update(props: props, staleDate: staleDate)
       }